+++ src/services/Elements.php
@@ -254,6 +254,12 @@ class Elements extends Component
             return null;
         }
 
+        if ($type === 'owldesign\\qarr\\elements\\Question') {
+            $type = 'questions';
+        } else {
+            $type = 'reviews';
+        }
+
         $table = '{{%qarr_' . $type . '}}';
 
         $result = Craft::$app->getDb()->createCommand()
@@ -431,6 +437,23 @@ class Elements extends Component
         if ($type == 'entry') {
             return 'craft\\elements\\Entry';
         }
+    }
+
+    public function markElementsAsDeletedByElementId($elementId, $date)
+    {
+        $reviews = Craft::$app->getDb()->createCommand()
+            ->update('{{%qarr_reviews}}',
+                ['dateDeleted' => $date->format('Y-m-d H:i:s')],
+                ['elementId' => $elementId])
+            ->execute();
+
+        $questions = Craft::$app->getDb()->createCommand()
+            ->update('{{%qarr_questions}}',
+                ['dateDeleted' => $date->format('Y-m-d H:i:s')],
+                ['elementId' => $elementId])
+            ->execute();
+
+        return true;
     }
 
     // Private Methods